Move "runserver" to Dockerfile
When deploying to OpenShift, the pod timed out waiting for something to happen.  I think it was waiting for the Django server to start, for which the Dockerfile didn't include a step.  Moving the "command" line from docker-compose.yml to a "CMD" line in Dockerfile worked well with docker-compose, so I hope the same will be the case when deploying to OpenShift.