prevent duplicate item note IDs
Item note IDs aren't referenced by other objects, but they must be unique.  I tried adding the last two digits of ROWNUM to the ID that was already being generated, but that made it too long to fit into INT.  I needed to change the destination table ID type to BIGINT.  That's not such a bad thing, but I didn't think it warranted a model change and migration.  So, I decided to take a chance and use only ROWNUM for the ID.  Since data will not be going back to the old DB, this seems like a reasonable choice.