#41 - sort items in correct order within topics
To put items in the correct order within topics, they should be sorted by item date, not simply year.  However, some older items do not have dates and it's not feasible to update their data.  So, in the case where an item doesn't have a date, use the start date of its volume instead.